src/agent_data_manager/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
faiss-cpu==1.11.0
fastapi==0.115.12
Flask>=2.3,<3.0
fonttools==4.57.0
freezegun==1.5.1
google-adk==0.3.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-firestore>=2.20.2,<2.21.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.12.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mcp==1.6.0
numpy==2.2.5
openai==1.76.0
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==25.0
passlib[bcrypt]==1.7.4
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf>=3.20.2,<6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
qdrant-client==1.12.1
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
retry>=0.9.2
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug>=2.3,<3.0
wrapt==1.17.2
zipp==3.21.0
